requirements.txt
google-generativeai
pytest
Flask
PyYAML
pdfkit
markdown2
accelerate